sum求和聚合函数中间累加器修改为BigDecimal实现精确计算

+++ aggregate-function-impl/src/test/java/com/yanggu/metric_calculate/core/aggregate_function/map/SortValueReturnKeyListMapAggregateFunctionTest.java
@@ -5,6 +5,7 @@ import com.yanggu.metric_calculate.core.aggregate_function.numeric.SumAggregateF
 import com.yanggu.metric_calculate.core.pojo.acc.MultiFieldData;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
@@ -31,18 +32,18 @@ class SortValueReturnKeyListMapAggregateFunctionTest {
 
     @Test
     void testGetResult() {
-        SortValueReturnKeyListMapAggregateFunction<Double, Double, Double> aggregateFunction = new SortValueReturnKeyListMapAggregateFunction<>();
+        SortValueReturnKeyListMapAggregateFunction<Double, BigDecimal, BigDecimal> aggregateFunction = new SortValueReturnKeyListMapAggregateFunction<>();
         aggregateFunction.setValueAggregateFunction(new SumAggregateFunction<>());
         //升序取4个
         aggregateFunction.setAsc(true);
         aggregateFunction.setLimit(4);
 
-        Map<MultiFieldData, Double> accumulator = new HashMap<>();
-        accumulator.put(create("test1"), 0.0D);
-        accumulator.put(create("test2"), -1.0D);
-        accumulator.put(create("test3"), 23.0D);
-        accumulator.put(create("test4"), 2.0D);
-        accumulator.put(create("test5"), 1.0D);
+        Map<MultiFieldData, BigDecimal> accumulator = new HashMap<>();
+        accumulator.put(create("test1"), BigDecimal.valueOf(0.0D));
+        accumulator.put(create("test2"), BigDecimal.valueOf(-1.0D));
+        accumulator.put(create("test3"), BigDecimal.valueOf(23.0D));
+        accumulator.put(create("test4"), BigDecimal.valueOf(2.0D));
+        accumulator.put(create("test5"), BigDecimal.valueOf(1.0D));
 
         List<List<Object>> result = aggregateFunction.getResult(accumulator);
         List<Object> list = result.stream().flatMap(Collection::stream).toList();
