+++ aggregate-function-impl/src/test/java/com/yanggu/metric_calculate/core/aggregate_function/map/SortValueReturnValueListMapAggregateFunctionTest.java
@@ -5,6 +5,7 @@ import com.yanggu.metric_calculate.core.aggregate_function.numeric.SumAggregateF
 import com.yanggu.metric_calculate.core.pojo.acc.MultiFieldData;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -33,20 +34,20 @@ class SortValueReturnValueListMapAggregateFunctionTest {
     @Test
     void testGetResult_Positive() {
         // 创建一个包含键值对的 HashMap
-        Map<MultiFieldData, Double> accumulator = new HashMap<>();
-        accumulator.put(create("key1"), 5.0D);
-        accumulator.put(create("key2"), 10.0D);
-        accumulator.put(create("key3"), 2.0D);
+        Map<MultiFieldData, BigDecimal> accumulator = new HashMap<>();
+        accumulator.put(create("key1"), BigDecimal.valueOf(5.0D));
+        accumulator.put(create("key2"), BigDecimal.valueOf(10.0));
+        accumulator.put(create("key3"), BigDecimal.valueOf(2.0D));
         // 创建一个 SortValueReturnValueListMapAggregateFunction 对象
-        SortValueReturnValueListMapAggregateFunction<Integer, Double, Double> sortMapAggFunc =
+        SortValueReturnValueListMapAggregateFunction<Integer, BigDecimal, BigDecimal> sortMapAggFunc =
                 new SortValueReturnValueListMapAggregateFunction<>();
         sortMapAggFunc.setValueAggregateFunction(new SumAggregateFunction<>());
         sortMapAggFunc.setLimit(2);
         sortMapAggFunc.setAsc(false);
         // 创建预期的整数类型的列表
-        List<Double> expectedList = new ArrayList<>();
-        expectedList.add(10.0D);
-        expectedList.add(5.0D);
+        List<BigDecimal> expectedList = new ArrayList<>();
+        expectedList.add(BigDecimal.valueOf(10.0D));
+        expectedList.add(BigDecimal.valueOf(5.0D));
         // 断言 SortValueReturnValueListMapAggregateFunction.getResult() 返回的值和预期的值一致
         assertEquals(expectedList, sortMapAggFunc.getResult(accumulator));
     }
@@ -54,22 +55,22 @@ class SortValueReturnValueListMapAggregateFunctionTest {
     @Test
     void testGetResult_WithSameValues_Positive() {
         // 创建一个包含键值对的 HashMap
-        Map<MultiFieldData, Double> accumulator = new HashMap<>();
-        accumulator.put(create("key1"), 5.0D);
-        accumulator.put(create("key2"), 10.0D);
-        accumulator.put(create("key3"), 2.0D);
-        accumulator.put(create("key4"), 5.0D);
+        Map<MultiFieldData, BigDecimal> accumulator = new HashMap<>();
+        accumulator.put(create("key1"), BigDecimal.valueOf(5.0D));
+        accumulator.put(create("key2"), BigDecimal.valueOf(10.0));
+        accumulator.put(create("key3"), BigDecimal.valueOf(2.0D));
+        accumulator.put(create("key4"), BigDecimal.valueOf(5.0D));
         // 创建一个 SortValueReturnValueListMapAggregateFunction 对象
-        SortValueReturnValueListMapAggregateFunction<Integer, Double, Double> sortMapAggFunc =
+        SortValueReturnValueListMapAggregateFunction<Integer, BigDecimal, BigDecimal> sortMapAggFunc =
                 new SortValueReturnValueListMapAggregateFunction<>();
         sortMapAggFunc.setValueAggregateFunction(new SumAggregateFunction<>());
         sortMapAggFunc.setLimit(3);
         sortMapAggFunc.setAsc(true);
         // 创建预期的整数类型的列表
-        List<Double> expectedList = new ArrayList<>();
-        expectedList.add(2.0D);
-        expectedList.add(5.0D);
-        expectedList.add(5.0D);
+        List<BigDecimal> expectedList = new ArrayList<>();
+        expectedList.add(BigDecimal.valueOf(2.0D));
+        expectedList.add(BigDecimal.valueOf(5.0D));
+        expectedList.add(BigDecimal.valueOf(5.0D));
         // 断言 SortValueReturnValueListMapAggregateFunction.getResult() 返回的值和预期的值一致
         assertEquals(expectedList, sortMapAggFunc.getResult(accumulator));
     }
