+++ aggregate-function-impl/src/test/java/com/yanggu/metric_calculate/core/aggregate_function/window/StateWindowAggregateFunctionTest.java
@@ -5,6 +5,8 @@ import org.dromara.hutool.core.lang.mutable.MutablePair;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
@@ -12,7 +14,7 @@ import static org.junit.jupiter.api.Assertions.*;
  */
 class StateWindowAggregateFunctionTest {
 
-    private StateWindowAggregateFunction<String, Integer, Double, Double> sumStateWindow;
+    private StateWindowAggregateFunction<String, Integer, BigDecimal, BigDecimal> sumStateWindow;
 
     @BeforeEach
     void init() {
@@ -22,37 +24,37 @@ class StateWindowAggregateFunctionTest {
 
     @Test
     void createAccumulator() {
-        MutablePair<String, Double> accumulator = sumStateWindow.createAccumulator();
+        MutablePair<String, BigDecimal> accumulator = sumStateWindow.createAccumulator();
         assertNotNull(accumulator);
         assertNull(accumulator.getLeft());
-        assertEquals(0.0D, accumulator.getRight(), 0.0D);
+        assertEquals(0.0D, accumulator.getRight().doubleValue(), 0.0D);
     }
 
     @Test
     void add1() {
-        MutablePair<String, Double> accumulator = sumStateWindow.createAccumulator();
+        MutablePair<String, BigDecimal> accumulator = sumStateWindow.createAccumulator();
         String key = "1";
         MutablePair<String, Integer> input = new MutablePair<>(key, 100);
-        MutablePair<String, Double> add = sumStateWindow.add(input, accumulator);
+        MutablePair<String, BigDecimal> add = sumStateWindow.add(input, accumulator);
         assertSame(add, accumulator);
         assertEquals(key, add.getLeft());
-        assertEquals(100.0D, add.getRight(), 0.0D);
+        assertEquals(100.0D, add.getRight().doubleValue(), 0.0D);
     }
 
     @Test
     void add2() {
-        MutablePair<String, Double> accumulator = sumStateWindow.createAccumulator();
+        MutablePair<String, BigDecimal> accumulator = sumStateWindow.createAccumulator();
         String key = "1";
         MutablePair<String, Integer> input = new MutablePair<>(key, 100);
-        MutablePair<String, Double> add = sumStateWindow.add(input, accumulator);
+        MutablePair<String, BigDecimal> add = sumStateWindow.add(input, accumulator);
         assertSame(add, accumulator);
         assertEquals(key, add.getLeft());
-        assertEquals(100.0D, add.getRight(), 0.0D);
+        assertEquals(100.0D, add.getRight().doubleValue(), 0.0D);
 
         add = sumStateWindow.add(input, add);
         assertSame(add, accumulator);
         assertEquals(key, add.getLeft());
-        assertEquals(200.0D, add.getRight(), 0.0D);
+        assertEquals(200.0D, add.getRight().doubleValue(), 0.0D);
 
         //当状态改变时, 累加器应该重新累加
         String key2 = "2";
@@ -60,18 +62,18 @@ class StateWindowAggregateFunctionTest {
         add = sumStateWindow.add(input2, add);
         assertSame(add, accumulator);
         assertEquals(key2, add.getLeft());
-        assertEquals(200.0D, add.getRight(), 0.0D);
+        assertEquals(200.0D, add.getRight().doubleValue(), 0.0D);
     }
 
     @Test
     void getResult() {
-        MutablePair<String, Double> accumulator = sumStateWindow.createAccumulator();
+        MutablePair<String, BigDecimal> accumulator = sumStateWindow.createAccumulator();
         String key = "1";
         MutablePair<String, Integer> input = new MutablePair<>(key, 100);
-        MutablePair<String, Double> add = sumStateWindow.add(input, accumulator);
+        MutablePair<String, BigDecimal> add = sumStateWindow.add(input, accumulator);
         add = sumStateWindow.add(input, add);
 
-        MutablePair<String, Double> result = sumStateWindow.getResult(add);
+        MutablePair<String, BigDecimal> result = sumStateWindow.getResult(add);
         System.out.println(result);
     }
