+++ metric-calculate-core/src/test/java/com/yanggu/metric_calculate/core/pojo/TestData.java
@@ -2,6 +2,7 @@ package com.yanggu.metric_calculate.core.pojo;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -10,14 +11,14 @@ public class TestData {
 
     private Input input;
 
-    private Map<List<String>, Double> output;
+    private Map<List<String>, BigDecimal> output;
 
     @Data
     public static class Input {
 
         private String accountNoIn;
 
-        private Double amount;
+        private BigDecimal amount;
     }
 
 }
